+++ .rspec
@@ -1,2 +1,3 @@
 --color
 --format progress
+--require spec_helper
