+++ lib/provisioner/cli/provision.rb
@@ -1,11 +1,14 @@
 require 'mixlib/cli'
-require 'pry'
-
 class Provisioner::CLI::Provision < Provisioner::CLI
   banner 'Usage: provisioner provision --config <path-to-config>.yml [options] '
 
-  def provisioner_command
-    Provisioner::Command::Provision.new(template_configuration.for_template(config[:template]), config)
+  def provisioner_command(type = 'knife')
+    command_class = type.capitalize
+    provisioner = Kernel::const_get("Provisioner::Command::#{command_class}")
+
+    @provisioner ||= provisioner.new(
+      template_configuration.for_template(config[:template]), config
+    )
   end
 
 end
