+++ spec/provisioner/command/provisioners/knife_spec.rb
@@ -1,11 +1,9 @@
-require 'spec_helper'
-
-describe Provisioner::Command::Provision do
+describe Provisioner::Command::Knife do
   let(:configuration) { Provisioner::Configuration.from_path('spec/fixtures/test.yml') }
 
   describe '#shell_commands' do
     context 'host is specified' do
-      let(:subject) { Provisioner::Command::Provision.new(template_configuration, number: '1') }
+      let(:subject) { Provisioner::Command::Knife.new(template_configuration, number: '1') }
 
       let(:command_options) { [
           '--image 9ec5c0c-a941-11e2-a7dc-57a6b041988f',
@@ -61,7 +59,7 @@ describe Provisioner::Command::Provision do
     end
 
     context 'host number is not specified' do
-      let(:subject) { Provisioner::Command::Provision.new(template_configuration) }
+      let(:subject) { Provisioner::Command::Knife.new(template_configuration) }
       let(:template_configuration) { configuration.for_template('memcached-sessions') }
 
       it 'returns an array of commands based on the configuration' do
